BAEL-315 added security to config and dsicovery servers.
