+++ spring-cloud/spring-cloud-bootstrap/config/src/main/java/com/baeldung/spring/cloud/bootstrap/config/SecurityConfig.java
@@ -1,8 +1,9 @@
-package com.baeldung.spring.cloud.integration.config;
+package com.baeldung.spring.cloud.bootstrap.config;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 
@@ -15,11 +16,24 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
         auth.inMemoryAuthentication()
                 .withUser("config_discUser")
                 .password("discPassword")
+                .roles("SYSTEM")
             .and()
                 .withUser("config_gatewayUser")
                 .password("gatewayPassword")
+                .roles("SYSTEM")
             .and()
                 .withUser("config_resourceUser")
-                .password("resourcePassword");
+                .password("resourcePassword")
+                .roles("SYSTEM");
+    }
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http
+                .authorizeRequests()
+                .anyRequest().authenticated()
+                .and()
+                .httpBasic().and()
+                .csrf().disable();
     }
 }
