+++ spring-cloud/spring-cloud-bootstrap/config/src/main/resources/application.properties
@@ -5,4 +5,4 @@ spring.cloud.config.server.git.uri=file:///${user.home}/application-config
 
 eureka.client.region = default
 eureka.client.registryFetchIntervalSeconds = 5
-eureka.client.serviceUrl.defaultZone=disc_configUser:configPassword@http://localhost:8082/eureka/
+eureka.client.serviceUrl.defaultZone=http://disc_configUser:configPassword@localhost:8082/eureka/
