+++ spring-cloud/spring-cloud-bootstrap/gateway/src/main/java/com/baeldung/spring/cloud/bootstrap/gateway/GatewayApplication.java
@@ -1,15 +1,15 @@
-package com.baeldung.spring.cloud.integration.config;
+package com.baeldung.spring.cloud.bootstrap.gateway;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.config.server.EnableConfigServer;
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
 
 @SpringBootApplication
-@EnableConfigServer
+@EnableZuulProxy
 @EnableEurekaClient
-public class ConfigApplication {
+public class GatewayApplication {
     public static void main(String[] args) {
-        SpringApplication.run(ConfigApplication.class, args);
+        SpringApplication.run(GatewayApplication.class, args);
     }
 }
