+++ spring-cloud/spring-cloud-bootstrap/gateway/src/main/resources/bootstrap.properties
@@ -4,4 +4,4 @@ spring.cloud.config.discovery.enabled=true
 spring.cloud.config.username=config_gatewayUser
 spring.cloud.config.password=gatewayPassword
 
-eureka.client.serviceUrl.defaultZone=http://localhost:8082/eureka/
+eureka.client.serviceUrl.defaultZone=http://disc_gatewayUser:gatewayPassword@localhost:8082/eureka/
