+++ spring-cloud/spring-cloud-bootstrap/resource/src/main/resources/bootstrap.properties
@@ -4,4 +4,4 @@ spring.cloud.config.discovery.enabled=true
 spring.cloud.config.username=config_resourceUser
 spring.cloud.config.password=resourcePassword
 
-eureka.client.serviceUrl.defaultZone=http://localhost:8082/eureka/
+eureka.client.serviceUrl.defaultZone=http://disc_resourceUser:resourcePassword@localhost:8082/eureka/
